Initial version of constrained Split Bregman
Former-commit-id: 8306029d5da01ad243729064f3400678d7e139e4

+++ apps/standalone/gpu/deblurring/2d/blur.cpp
@@ -14,7 +14,7 @@
 using namespace std;
 
 // Define desired precision (note that decent deblurring of noisy images requires double precision)
-typedef double _real; 
+typedef float _real; 
 typedef complext<_real>::Type _complext;
 
 int main( int argc, char** argv)
